Changed cigar string to a primarily list based method. All three versions of neat_cigar are present in this commit, in case I need to walk back
